keiba-v2/pyproject.toml
[project]
name = "keiba-v2"
version = "4.0.0"
description = "KeibaCICD v4 - JRA-VAN Native Architecture"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]

[tool.ruff]
line-length = 100
